requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.11
Django==4.0.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
flake8==5.0.2
flake8-html==0.4.2
idna==3.3
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
psycopg2==2.9.3
pycodestyle==2.9.0
pyflakes==2.5.0
Pygments==2.12.0
PyJWT==2.3.0
python-dotenv==0.20.0
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
